@@ -1,4 +1,4 @@
-@page "{userEmail}/{testId:guid}"
+@page "{userEmail}/{testId}"
 @model presentation.Pages.Admin.Test.DetailModel
 
 
