@@ -1,4 +1,4 @@
-@page "{id:guid}"
+@page "{id}"
 @model presentation.Pages.Submissions.DetailModel
 @using presentation.Services;
 
@@ -59,5 +59,5 @@ else
 {
     @await Component.InvokeAsync(
         "TestList",
-        new { userEmail = User.GetEmail(), taskId = "", submissionId = Model.MySubmission.Id })
+        new { userEmail = User.GetEmail(), submissionId = Model.MySubmission.Id })
 }
