@@ -1,4 +1,4 @@
-@page "{guid:guid}"
+@page "{id}"
 @model presentation.Pages.Tests.DetailModel
 @{
 }
