@@ -1,11 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using submissions.Data;
+using MongoDB.Bson;
 using submissions.Models;
 using submissions.Services;
 
@@ -15,114 +14,157 @@ namespace submissions.Controllers
     [Route("[controller]")]
     public class SubmissionController : Controller
     {
-        public SubmissionController(CosmosContext context, ILogger<SubmissionController> logger, StorageService storage)
+        public SubmissionController(
+            SubmissionsService submissions,
+            ILogger<SubmissionController> logger,
+            StorageService storage)
         {
-            _context = context;
+            _submissions = submissions;
             _logger = logger;
             _storage = storage;
         }
 
         [HttpGet]
-        public IEnumerable<Submission> Index()
+        public async Task<ActionResult<List<Submission>>> OnGetAsync()
         {
-            return _context.Submissions.OrderByDescending(submission => submission.SubmittedOn).Take(30);
+            try
+            {
+                _logger.LogTrace("Retrieving all submissions");
+
+                return Ok(await _submissions.GetAsync(30));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError("Error during retrieval of data. {e}", e);
+                return StatusCode(500);  // Internal error
+            }
         }
 
         [HttpGet("{email}")]
-        public ActionResult<IAsyncEnumerable<Submission>> IndexForUser(string email, [FromQuery]Guid? taskId)
+        public async Task<ActionResult<List<Submission>>> OnGetAsync(string email, [FromQuery]string taskId)
         {
-            var queryable = _context.Submissions.AsQueryable();
-
-            if (taskId != null)
+            try
             {
-                queryable = queryable.Where(
-                    test => test.TaskId == taskId.ToString());
-            }
+                _logger.LogTrace("Retrieving all submissions for user {email} with task id {taskId}", email, taskId);
 
-            return Ok(queryable.OrderByDescending(test => test.SubmittedOn).Take(30).AsAsyncEnumerable());
+                return Ok(await _submissions.GetForUserAsync(email, taskId, 30));
+            }
+            catch (System.FormatException e)
+            {
+                _logger.LogError("Wrong input format. {e}", e);
+                return StatusCode(400);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError("Error during retrieval of data. {e}", e);
+                return StatusCode(500);  // Internal error
+            }
         }
 
 
         // https://www.yogihosting.com/aspnet-core-api-controllers/
 
         [HttpGet("{email}/{submissionId}")]
-        public async Task<ActionResult<SubmissionRest>> GetSubmissionAsync(
+        public async Task<ActionResult<Submission>> GetSubmissionAsync(
             string email, string submissionId, [FromQuery]string contentFormat)
         {
-            _logger.LogTrace("Retrieving submission {id} for {email}", submissionId, email);
-            Submission result = await _context.Submissions.FirstOrDefaultAsync(
-                s => s.UserEmail == email && s.Id == submissionId);
-            if (result == null)
-            {
-                _logger.LogTrace("Submission not found");
-                return NotFound();
-            }
             try
             {
-                string content;
-                List<string> bloblPath = new() { email, submissionId };
-                _logger.LogTrace("Found in db, retrieving blob");
+                _logger.LogTrace("Retrieving submission {id} for {email}", submissionId, email);
+                Submission result = await _submissions.GetAsync(submissionId);
+
+                if (result.UserEmail != email)
+                {
+                    _logger.LogWarning("Submission email isn't right");
+                    return NotFound();
+                }
+
+                _logger.LogTrace("Submission found in db, retrieving blob");
+
+                List<string> blobPath = new() { email, submissionId };
+                    
                 if (contentFormat == "url")
                 {
                     _logger.LogTrace("Retrieving blob as sas url");
-                    content = _storage.GetUrlBlob(bloblPath);
+                    result.Content = _storage.GetUrlBlob(blobPath);
                 }
                 else
                 {
                     _logger.LogTrace("Retrieving blob as binary data");
-                    content = (await _storage.DownloadBlobAsync(bloblPath)).ToString();
+                    result.Content = (await _storage.DownloadBlobAsync(blobPath)).ToString();
                 }
+
                 _logger.LogTrace("Retrieved blob data successfuly");
-                return Ok(new SubmissionRest(result, content));
+                return Ok(result);
             }
-            catch(Exception e)
+            catch (System.FormatException e)
             {
-                _logger.LogError("Error during blob retrieval {e}", e);
+                _logger.LogError("Wrong input format. {e}", e);
+                return StatusCode(400);
+            }
+            catch (System.InvalidOperationException e) // this is the exception for not found
+            {
+                _logger.LogError("Submission id not found. {e}", e);
+                return NotFound();
+            }
+            catch (Exception e)
+            {
+                _logger.LogError("Error during data retrieval. {e}", e);
                 return StatusCode(500);  // Internal error
             }
         }
 
         [HttpPost]
-        public async Task<ActionResult<Submission>> Create([FromBody] SubmissionRest submissionData)
+        public async Task<ActionResult<Submission>> Create([FromBody] Submission submission)
         {
-            if (submissionData.UserEmail != StorageService.NormalizeFileName(submissionData.UserEmail))
+            if (submission.UserEmail != StorageService.NormalizeFileName(submission.UserEmail))
             {
                 _logger.LogWarning(
-                    "Rejecting submission create, due to invalid email {email}", submissionData.UserEmail);
+                    "Rejecting submission create, due to invalid email {email}", submission.UserEmail);
                 return BadRequest();
             }
-            _logger.LogTrace("Saving submission {submission}", submissionData);
-            var submission = submissionData.GenerateSubmission();
-            _context.Submissions.Add(submission);
-            await _context.SaveChangesAsync();
-            _logger.LogInformation("Submission saved to cosmos, now to blob");
+
+            if (submission.Content == null)
+            {
+                _logger.LogWarning("Content cannot be empty on create.");
+                return BadRequest();
+            }
+
+            if (!ObjectId.TryParse(submission.TaskId, out _))
+            {
+                _logger.LogWarning("TaskId must be valid ObjectId.");
+                return BadRequest();
+            }
+
+            _logger.LogTrace("Saving submission {obj}", JsonSerializer.Serialize(submission));
+
+            await _submissions.CreateAsync(submission);
+
+            _logger.LogInformation("Submission saved to mongo, now to blob");
 
             try
             {
                 await _storage.UploadBlobAsync(
                     new List<string> { submission.UserEmail, submission.Id.ToString() },
-                    new BinaryData(submissionData.Content)
+                    new BinaryData(submission.Content)
                     );
+
                 return Created("N/A", submission);
             }
-            catch(Exception e)
+            catch (Exception e)
             {
                 // TODO: Test if this deletion really works
-                _logger.LogWarning("Unable to save to blob, reverting db {e}", e);
-                _context.Submissions.Remove(submission);
-                await _context.SaveChangesAsync();
+                _logger.LogWarning("Unable to save to blob, reverting db. {e}", e);
+                await _submissions.RemoveAsync(submission.Id);
+
                 // If there is exception, let it be some error
-                _logger.LogTrace("Reverted  db succesfully");
+                _logger.LogTrace("Reverted db succesfully");
                 return StatusCode(500);
             }
         }
 
-
-
-
-        // POST: SubmissionController/Delete/5
-        private CosmosContext _context = null;
-        private ILogger<SubmissionController> _logger = null;
-        private StorageService _storage = null;
+        private readonly SubmissionsService _submissions = null;
+        private readonly ILogger<SubmissionController> _logger = null;
+        private readonly StorageService _storage = null;
     }
 }
